endpoint for fight with monsters

+++ ai-service/src/main/java/com/codecool/aiservice/model/Roll.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class Roll{
-    private int result;
+    private Long result;
     private String dice;
 }
